+++ apps/web/.env.example
@@ -10,7 +10,7 @@ NEXT_PUBLIC_APP_SHORT_DOMAIN=dub.sh
 # Get your Tinybird Auth Token here: https://www.tinybird.co/docs/concepts/auth-tokens.html
 TINYBIRD_API_KEY=
 # Varies based on your Tinybird region: https://www.tinybird.co/docs/api-reference/api-reference.html#regions-and-endpoints
-TINYBIRD_API_URL=https://api.us-east.tinybird.co
+TINYBIRD_API_URL=https://api.tinybird.co
 
 # Vercel's Domains API – required for adding and removing domains
 # Learn how to set this up: https://vercel.com/templates/next.js/domains-api
