+++ packages/utils/src/constants/domains.ts
@@ -1,0 +1,21 @@
+export const SECOND_LEVEL_DOMAINS = new Set([
+  "com",
+  "co",
+  "net",
+  "org",
+  "edu",
+  "gov",
+  "in",
+]);
+
+export const SPECIAL_APEX_DOMAINS = new Set([
+  "my.id",
+  "github.io",
+  "vercel.app",
+  "now.sh",
+  "pages.dev",
+  "webflow.io",
+  "netlify.app",
+  "fly.dev",
+  "web.app",
+]);
