+++ packages/utils/src/constants/layout.ts
@@ -1,0 +1,16 @@
+export const HIDE_BACKGROUND_SEGMENTS = [
+  "stats",
+  "changelog",
+  "terms",
+  "privacy",
+  "abuse",
+];
+
+export const ALL_TOOLS = [
+  { name: "ChatGPT Link Shortener", slug: "chatgpt-link-shortener" },
+  { name: "Amazon Link Shortener", slug: "amazon-link-shortener" },
+  { name: "Spotify Link Shortener", slug: "spotify-link-shortener" },
+  { name: "Metatags API", slug: "metatags" },
+  { name: "QR Code API", slug: "qr-code" },
+  { name: "Link Inspector", slug: "inspector" },
+];
