+++ packages/utils/src/constants/localhost.ts
@@ -1,0 +1,8 @@
+export const LOCALHOST_GEO_DATA = {
+  city: "San Francisco",
+  region: "CA",
+  country: "US",
+  latitude: "37.7695",
+  longitude: "-122.385",
+};
+export const LOCALHOST_IP = "63.141.57.109";
