+++ packages/utils/src/constants/middleware.ts
@@ -1,0 +1,20 @@
+export const DEFAULT_REDIRECTS = {
+  home: "https://dub.co",
+  dub: "https://dub.co",
+  signin: "https://app.dub.co/login",
+  login: "https://app.dub.co/login",
+  register: "https://app.dub.co/register",
+  signup: "https://app.dub.co/register",
+  app: "https://app.dub.co",
+  dashboard: "https://app.dub.co",
+  links: "https://app.dub.co/links",
+  settings: "https://app.dub.co/settings",
+  welcome: "https://app.dub.co/welcome",
+  discord: "https://twitter.com/dubdotco", // placeholder for now
+};
+
+export const DUB_HEADERS = {
+  headers: {
+    "x-powered-by": "Dub.co - Link management for modern marketing teams",
+  },
+};
