+++ packages/utils/src/constants/misc.ts
@@ -1,0 +1,25 @@
+export const DEFAULT_LINK_PROPS = {
+  key: "",
+  url: "",
+  domain: "",
+  archived: false,
+  expiresAt: null,
+  password: null,
+
+  title: null,
+  description: null,
+  image: null,
+  rewrite: false,
+  ios: null,
+  android: null,
+
+  clicks: 0,
+  userId: "",
+
+  proxy: false,
+};
+
+export const GOOGLE_FAVICON_URL =
+  "https://www.google.com/s2/favicons?sz=64&domain_url=";
+
+export const PAGINATION_LIMIT = 100;
